# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Example API",
        "contact": {}
    },
    "host": "petstore.swagger.io",
    "basePath": "/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "登陆",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "登陆用户名和密码",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LoginParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/bookType/add": {
            "post": {
                "description": "新增📖类型",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "新增📖的数据",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modal.BookType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/bookType/delete": {
            "delete": {
                "description": "删除图书分类",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "删除图书分类的🆔",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/bookType/list": {
            "post": {
                "description": "获取图书分类列表",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modal.BookTypeReqParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.TableResult"
                        }
                    }
                }
            }
        },
        "/v1/bookType/treeList": {
            "get": {
                "description": "返回树形结构的图书类型",
                "consumes": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/modal.TreeBookType"
                            }
                        }
                    }
                }
            }
        },
        "/v1/bookType/update": {
            "post": {
                "description": "更新图书分类列表",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modal.BookType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/role/add": {
            "post": {
                "description": "新增角色",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "新增角色的数据",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modal.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/role/delete": {
            "delete": {
                "description": "删除角色",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "删除角色的id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/role/list": {
            "post": {
                "description": "查询角色",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "查询角色的信息",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/modal.RoleParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.RoleListResult"
                        }
                    }
                }
            }
        },
        "/v1/role/option": {
            "get": {
                "description": "查询角色的下拉选项",
                "consumes": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.SelectOption"
                        }
                    }
                }
            }
        },
        "/v1/role/update": {
            "post": {
                "description": "修改角色",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "新增角色的数据",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modal.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/user/add": {
            "post": {
                "description": "新增用户",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "新增用户的数据",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modal.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/user/delete": {
            "delete": {
                "description": "删除用户",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "删除用户",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        },
        "/v1/user/list": {
            "post": {
                "description": "查询用户",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "查询用户的信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modal.QueryUserParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.UserListResult"
                        }
                    }
                }
            }
        },
        "/v1/user/update": {
            "post": {
                "description": "修改用户",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "修改用户的数据",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modal.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modal.Result"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.LoginParams": {
            "type": "object",
            "properties": {
                "user_name": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                }
            }
        },
        "modal.BookType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "string"
                },
                "pId": {
                    "type": "integer"
                },
                "remake": {
                    "type": "string"
                },
                "typeName": {
                    "type": "string"
                }
            }
        },
        "modal.BookTypeReqParams": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "keyWord": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "order_by": {
                    "type": "string"
                },
                "order_type": {
                    "type": "string"
                },
                "pName": {
                    "type": "string"
                },
                "pageSize": {
                    "type": "integer"
                },
                "typeName": {
                    "type": "string"
                }
            }
        },
        "modal.QueryUserParams": {
            "type": "object",
            "properties": {
                "borrow_book_count": {
                    "type": "string"
                },
                "current": {
                    "type": "integer"
                },
                "order_by": {
                    "type": "string"
                },
                "order_type": {
                    "type": "string"
                },
                "pageSize": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "modal.Result": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "modal.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                },
                "role_weight": {
                    "type": "integer"
                }
            }
        },
        "modal.RoleListResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                },
                "role_weight": {
                    "type": "string"
                }
            }
        },
        "modal.RoleParams": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "modal.SelectOption": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "modal.TableResult": {
            "type": "object",
            "properties": {
                "data": {},
                "total": {
                    "type": "integer"
                }
            }
        },
        "modal.TreeBookType": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modal.TreeBookType"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pId": {
                    "type": "integer"
                },
                "remake": {
                    "type": "string"
                }
            }
        },
        "modal.User": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "borrow_book_count": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "remake": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                }
            }
        },
        "modal.UserListResult": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "borrow_book_count": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "remake": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        }
    }
}